Contain all item manipulation behaviour in ItemManager. Delete ItemAppender module.
