--- app/models/spree/admin/item_appender.rb
@@ -1,11 +1,0 @@
-module Spree
-  module Admin
-    module ItemAppender
-      def add(item)
-        raise KeyError, "Item with key #{item.key} already exists" if index_for_key(item.key)
-
-        @items << item
-      end
-    end
-  end
-end
