+++ app/models/spree/admin/item_manager.rb
@@ -1,6 +1,12 @@
 module Spree
   module Admin
     module ItemManager
+      def add(item)
+        raise KeyError, "Item with key #{item.key} already exists" if index_for_key(item.key)
+
+        @items << item
+      end
+
       def child_with_key?(key)
         index_for_key(key).present?
       end
