+++ spec/models/spree/admin/main_menu/root_spec.rb
@@ -11,7 +11,6 @@ module Spree
       end
 
       it_behaves_like "implements item manipulation and query methods"
-      it_behaves_like "implements item appendage"
 
       describe '#key' do
         subject { root.key }
