+++ spec/support/admin/manipulation_and_query_methods.rb
@@ -1,4 +1,25 @@
 RSpec.shared_examples "implements item manipulation and query methods" do
+  describe '#add' do
+    subject { root.add(item) }
+    let(:item) { double(key: 'test') }
+
+    context "when there's no item with a particular key" do
+
+      it 'appends an item' do
+        subject
+        expect(root.items).to include(item)
+      end
+    end
+
+    context 'when there is an item with a particular key' do
+      let(:items) { [item] }
+
+      it 'raises an error' do
+        expect { subject }.to raise_error(KeyError)
+      end
+    end
+  end
+
   describe '#child_with_key?' do
     subject { root.child_with_key?(key) }
     let(:key) { 'key' }
